Moved manager ish stuff from ghedesigner/ghe/manager.py to just ghedesigner/main.py
